added MediUX support, added tpdb user support
